# docs/services/openapi-facilitator.json
{
  "components": {
    "schemas": {
      "ChatRequest": {
        "description": "One user message to the AICC facilitator (P4).",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "cds_commit_args": {
        "additionalProperties": false,
        "properties": {
          "include_unverified": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Include Unverified"
          }
        },
        "title": "cds_commit_args",
        "type": "object"
      },
...
      "cds_queue_add_args": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "description": "The content statement (for a position: the stance rationale).",
            "title": "Description",
            "type": "string"
          },
          "question": {
            "description": "The open question to resolve later.",
            "title": "Question",
            "type": "string"
          },
          "slug": {
            "description": "Short kebab-case id for this record (e.g. 'reach-a-human').",
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "question"
        ],
        "title": "cds_queue_add_args",
        "type": "object"
      },
...
      "cds_synthesis_args": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "description": "The content statement (for a position: the stance rationale).",
            "title": "Description",
            "type": "string"
          },
          "slug": {
            "description": "Short kebab-case id for this record (e.g. 'reach-a-human').",
            "title": "Slug",
            "type": "string"
          },
          "title": {
            "description": "Human title of the mapping.",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "title"
        ],
        "title": "cds_synthesis_args",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Guided authoring for a concept-definition record. Each /tools endpoint checks its input, writes a draft into your session (never the shared record), and verification gives advisory feedback while you compose. Publishing a change to the shared record is a separate, human-approved commit step.",
    "title": "cds facilitator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/chat": {
      "post": {
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Chat Chat Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chat"
      }
    },
...
    "/tools/cds_commit": {
      "post": {
        "description": "Merge staging into the canonical record (requires the cds-reviewer role bound at server start; publishing is a human decision); returns the executed change plan.",
        "operationId": "cds_commit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cds_commit_args"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cds Commit"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cds Commit"
      }
    },
...
    "/tools/cds_queue_add": {
      "post": {
        "description": "File a retrieval item for an unsecured source: authoring on it stops until a human secures it.",
        "operationId": "cds_queue_add",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cds_queue_add_args"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cds Queue Add"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cds Queue Add"
      }
    },
...
    "/tools/cds_synthesis": {
      "post": {
        "description": "Create/update the Synthesis (candidate into staging).",
        "operationId": "cds_synthesis",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/cds_synthesis_args"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cds Synthesis"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cds Synthesis"
      }
    },
...
  }
}